Refactored SnowMound Using Inheritance
Modified the snowmound logic to utilize inheritance by taking functionality from a parent blueprint called BP_InteractableBase. The logic behind updating it to this methodology is that I will use the parent blueprint for the shop interaction system and also the firepit system. Upon going in range of an interactable object a textprompt will also display infront of the player (adjusted from the original print string that was appearing).